Split `java.time` tests out of `StopwatchTest`.
RELNOTES=n/a
PiperOrigin-RevId: 376890185
