+++ guava-tests/test/com/google/common/base/StopwatchTest.java
@@ -21,9 +21,7 @@ import static java.util.concurrent.TimeUnit.MILLISECONDS;
 import static java.util.concurrent.TimeUnit.NANOSECONDS;
 
 import com.google.common.annotations.GwtCompatible;
-import com.google.common.annotations.GwtIncompatible;
 import com.google.common.testing.FakeTicker;
-import java.time.Duration;
 import junit.framework.TestCase;
 
 /**
@@ -166,15 +164,6 @@ public class StopwatchTest extends TestCase {
     assertEquals(0, stopwatch.elapsed(MILLISECONDS));
     ticker.advance(1);
     assertEquals(1, stopwatch.elapsed(MILLISECONDS));
-  }
-
-  @GwtIncompatible
-  public void testElapsed_duration() {
-    stopwatch.start();
-    ticker.advance(999999);
-    assertEquals(Duration.ofNanos(999999), stopwatch.elapsed());
-    ticker.advance(1);
-    assertEquals(Duration.ofMillis(1), stopwatch.elapsed());
   }
 
   public void testToString() {
